Use 64-bit type for length
This change allow the max length to be UINT64_MAX.

The test vectors have been updated with the hash of a message of 10 GB.
Add -DTEST_VECTORS_LONG to the Makefile to enable the test of this
vector.

+++ sha2.h
@@ -54,15 +54,15 @@ extern "C" {
 #endif
 
 typedef struct {
-    uint32 tot_len;
-    uint32 len;
+    uint64 tot_len;
+    uint64 len;
     uint8 block[2 * SHA256_BLOCK_SIZE];
     uint32 h[8];
 } sha256_ctx;
 
 typedef struct {
-    uint32 tot_len;
-    uint32 len;
+    uint64 tot_len;
+    uint64 len;
     uint8 block[2 * SHA512_BLOCK_SIZE];
     uint64 h[8];
 } sha512_ctx;
@@ -71,24 +71,24 @@ typedef sha512_ctx sha384_ctx;
 typedef sha256_ctx sha224_ctx;
 
 void sha224_init(sha224_ctx *ctx);
-void sha224_update(sha224_ctx *ctx, const uint8 *message, uint32 len);
+void sha224_update(sha224_ctx *ctx, const uint8 *message, uint64 len);
 void sha224_final(sha224_ctx *ctx, uint8 *digest);
-void sha224(const uint8 *message, uint32 len, uint8 *digest);
+void sha224(const uint8 *message, uint64 len, uint8 *digest);
 
 void sha256_init(sha256_ctx * ctx);
-void sha256_update(sha256_ctx *ctx, const uint8 *message, uint32 len);
+void sha256_update(sha256_ctx *ctx, const uint8 *message, uint64 len);
 void sha256_final(sha256_ctx *ctx, uint8 *digest);
-void sha256(const uint8 *message, uint32 len, uint8 *digest);
+void sha256(const uint8 *message, uint64 len, uint8 *digest);
 
 void sha384_init(sha384_ctx *ctx);
-void sha384_update(sha384_ctx *ctx, const uint8 *message, uint32 len);
+void sha384_update(sha384_ctx *ctx, const uint8 *message, uint64 len);
 void sha384_final(sha384_ctx *ctx, uint8 *digest);
-void sha384(const uint8 *message, uint32 len, uint8 *digest);
+void sha384(const uint8 *message, uint64 len, uint8 *digest);
 
 void sha512_init(sha512_ctx *ctx);
-void sha512_update(sha512_ctx *ctx, const uint8 *message, uint32 len);
+void sha512_update(sha512_ctx *ctx, const uint8 *message, uint64 len);
 void sha512_final(sha512_ctx *ctx, uint8 *digest);
-void sha512(const uint8 *message, uint32 len, uint8 *digest);
+void sha512(const uint8 *message, uint64 len, uint8 *digest);
 
 #ifdef __cplusplus
 }
